Add files via upload
dditional Description
Sample Data Usage
Please note that this project utilizes sample data for demonstration purposes. The sample data consists of approximately 27,000 records, which have been extracted from a publicly available dataset using web scraping techniques.

Data Representation
The sample data includes various trade metrics such as product codes, labels, transaction values, countries, years, import/export indicators, etc. These metrics are representative of the types of information commonly found in trade datasets.

Scalability and Customization
While this project focuses on a specific dataset and sample size, the techniques and methodologies employed can be scaled and customized to handle larger datasets or different sources of trade data. Users are encouraged to modify the code and dashboard to suit their specific requirements and datasets.

Performance Considerations
Due to the limited size of the sample data, the performance of the web scraping process and data cleaning operations may differ when applied to larger datasets. It is recommended to assess performance considerations and optimize the code as needed when working with larger volumes of data.

Contribution and Collaboration
Contributions and feedback are welcome! If you have suggestions for improving the code, adding new features, or enhancing the dashboard, feel free to submit a pull request or open an issue. Collaboration in refining and extending the functionality of this project is encouraged.